refactor: 인수 테스트 격리·기동 인프라 정비 - 글로벌 정리 훅·seed-data 로 시나리오 격리, 컨테이너 reuse 로 부팅 비용 제거
