+++ api/src/test/java/ksh/tryptobackend/acceptance/steps/batch/BatchOrchestrationStepDefinition.java
@@ -1,4 +1,4 @@
-package ksh.tryptobackend.acceptance.steps;
+package ksh.tryptobackend.acceptance.steps.batch;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -136,11 +136,14 @@ public class BatchOrchestrationStepDefinition {
     public void 오케스트레이션용_랭킹_대상_라운드가_존재한다() {
         for (int i = 0; i < 5; i++) {
             jdbcTemplate.update(
-                    "INSERT INTO orders (idempotency_key, wallet_id, exchange_coin_id, order_type,"
+                    "INSERT INTO orders (idempotency_key, user_id, wallet_id, exchange_coin_id,"
+                            + " coin_id, base_coin_id, exchange_name, market_symbol, order_type,"
                             + " side, order_amount, quantity, price, filled_price, fee, fee_rate,"
-                            + " status, created_at, filled_at) VALUES (?, ?, 1, 'MARKET', 'BUY',"
-                            + " 100000, 0.001, 50000000, 50000000, 50, 0.0005, 'FILLED', ?, ?)",
+                            + " status, created_at, filled_at) VALUES (?, ?, ?, 10, 2, 1, 'UPBIT',"
+                            + " 'BTC/KRW', 'MARKET', 'BUY', 100000, 0.001, 50000000, 50000000, 50,"
+                            + " 0.0005, 'FILLED', ?, ?)",
                     java.util.UUID.randomUUID().toString(),
+                    USER_ID,
                     WALLET_ID,
                     LocalDateTime.now(),
                     LocalDateTime.now());
