+++ api/src/test/java/ksh/tryptobackend/acceptance/steps/investmentround/EmergencyFundingStepDefinition.java
@@ -1,9 +1,8 @@
-package ksh.tryptobackend.acceptance.steps;
+package ksh.tryptobackend.acceptance.steps.investmentround;
 
 import static ksh.tryptobackend.marketdata.domain.model.ExchangeMarketType.DOMESTIC;
 
 import com.jayway.jsonpath.JsonPath;
-import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
@@ -56,27 +55,8 @@ public class EmergencyFundingStepDefinition {
         this.walletBalanceJpaRepository = walletBalanceJpaRepository;
     }
 
-    @Before("@emergency-funding")
-    public void setUp() {
-        emergencyFundingJpaRepository.deleteAllInBatch();
-        investmentRuleJpaRepository.deleteAllInBatch();
-        walletBalanceJpaRepository.deleteAllInBatch();
-        walletJpaRepository.deleteAllInBatch();
-        investmentRoundJpaRepository.deleteAllInBatch();
-        exchangeJpaRepository.deleteAllInBatch();
-        activeRoundId = null;
-        disabledFundingRoundId = null;
-        lastIdempotencyKey = null;
-    }
-
     @Given("긴급자금용 거래소 메타데이터가 준비되어 있다")
-    public void 긴급자금용_거래소_메타데이터가_준비되어_있다() {
-        exchangeJpaRepository.saveAll(
-                List.of(
-                        new ExchangeJpaEntity(1L, "UPBIT", DOMESTIC, 1L, new BigDecimal("0.0005")),
-                        new ExchangeJpaEntity(
-                                2L, "BITHUMB", DOMESTIC, 1L, new BigDecimal("0.0005"))));
-    }
+    public void 긴급자금용_거래소_메타데이터가_준비되어_있다() {}
 
     @Given("활성 라운드가 존재한다")
     public void 활성_라운드가_존재한다() {
