+++ api/src/test/java/ksh/tryptobackend/acceptance/steps/ranking/RankingStatsStepDefinition.java
@@ -1,6 +1,5 @@
-package ksh.tryptobackend.acceptance.steps;
+package ksh.tryptobackend.acceptance.steps.ranking;
 
-import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
@@ -18,11 +17,6 @@ public class RankingStatsStepDefinition {
     public RankingStatsStepDefinition(CommonApiClient apiClient, JdbcTemplate jdbcTemplate) {
         this.apiClient = apiClient;
         this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Before
-    public void setUp() {
-        jdbcTemplate.execute("DELETE FROM ranking");
     }
 
     @Given("랭킹 통계 테스트 데이터가 준비되어 있다")
