+++ api/src/test/java/ksh/tryptobackend/acceptance/steps/user/ChangePortfolioVisibilityStepDefinition.java
@@ -1,6 +1,5 @@
-package ksh.tryptobackend.acceptance.steps;
+package ksh.tryptobackend.acceptance.steps.user;
 
-import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
@@ -18,11 +17,6 @@ public class ChangePortfolioVisibilityStepDefinition {
             CommonApiClient apiClient, JdbcTemplate jdbcTemplate) {
         this.apiClient = apiClient;
         this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Before
-    public void setUp() {
-        jdbcTemplate.execute("DELETE FROM user WHERE user_id IN (1, 2)");
     }
 
     @Given("포트폴리오 공개 설정 테스트 데이터가 준비되어 있다")
