+++ api/src/test/java/ksh/tryptobackend/acceptance/steps/wallet/WalletAssetsStepDefinition.java
@@ -1,116 +1,111 @@
-package ksh.tryptobackend.acceptance.steps;
+package ksh.tryptobackend.acceptance.steps.wallet;
 
-import io.cucumber.java.Before;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import ksh.tryptobackend.acceptance.mock.MockHoldingAdapter;
-import ksh.tryptobackend.acceptance.mock.MockLivePriceAdapter;
 import ksh.tryptobackend.acceptance.testclient.CommonApiClient;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-public class MyHoldingsStepDefinition {
+public class WalletAssetsStepDefinition {
 
-    private static final Long EXCHANGE_ID = 1L;
+    private static final Long EXCHANGE_UPBIT_ID = 1L;
+    private static final Long EXCHANGE_BITHUMB_ID = 2L;
     private static final Long KRW_COIN_ID = 1L;
     private static final Long BTC_COIN_ID = 2L;
     private static final Long ETH_COIN_ID = 3L;
-    private static final Long BTC_EXCHANGE_COIN_ID = 10L;
-    private static final Long ETH_EXCHANGE_COIN_ID = 11L;
 
     private final CommonApiClient apiClient;
     private final JdbcTemplate jdbcTemplate;
-    private final MockHoldingAdapter holdingAdapter;
-    private final MockLivePriceAdapter livePriceAdapter;
 
-    public MyHoldingsStepDefinition(
-            CommonApiClient apiClient,
-            JdbcTemplate jdbcTemplate,
-            MockHoldingAdapter holdingAdapter,
-            MockLivePriceAdapter livePriceAdapter) {
+    public WalletAssetsStepDefinition(CommonApiClient apiClient, JdbcTemplate jdbcTemplate) {
         this.apiClient = apiClient;
         this.jdbcTemplate = jdbcTemplate;
-        this.holdingAdapter = holdingAdapter;
-        this.livePriceAdapter = livePriceAdapter;
     }
 
-    @Before
-    public void setUp() {
-        holdingAdapter.clear();
-        livePriceAdapter.clear();
-    }
-
-    @Given("보유 현황 테스트 데이터가 준비되어 있다")
-    public void 보유_현황_테스트_데이터가_준비되어_있다() {
+    @Given("잔고 조회 테스트 데이터가 준비되어 있다")
+    public void 잔고_조회_테스트_데이터가_준비되어_있다() {
         insertUsers();
-        insertExchangeAndCoins();
-        insertExchangeCoinMappings();
-        insertInvestmentRound();
+        insertCoins();
+        insertExchanges();
+        insertExchangeCoins();
+        insertInvestmentRounds();
         insertWallets();
         insertWalletBalances();
-        setUpHoldings();
-        setUpLivePrices();
     }
 
-    @When("유저 {long}이 지갑 {long}의 포트폴리오를 조회한다")
-    public void 유저가_지갑의_포트폴리오를_조회한다(Long userId, Long walletId) {
-        apiClient.get("/api/users/" + userId + "/wallets/" + walletId + "/portfolio");
+    @When("유저 {long}이 지갑 {long}의 잔고를 조회한다")
+    public void 유저가_지갑의_잔고를_조회한다(Long userId, Long walletId) {
+        apiClient.get("/api/users/" + userId + "/wallets/" + walletId + "/balances");
     }
 
-    @Then("거래소 ID는 {long}이다")
-    public void 거래소_ID는_이다(Long exchangeId) {
+    @Then("기축통화 사용 가능 잔고는 {bigdecimal}이다")
+    public void 기축통화_사용_가능_잔고는_이다(BigDecimal amount) {
         apiClient
                 .getLastResponse()
                 .expectBody()
-                .jsonPath("$.data.exchangeId")
-                .isEqualTo(exchangeId);
+                .jsonPath("$.data.baseCurrencyAvailable")
+                .value(
+                        value ->
+                                assertThat(new BigDecimal(value.toString()).compareTo(amount))
+                                        .isZero());
     }
 
-    @Then("기축통화 심볼은 {string}이다")
-    public void 기축통화_심볼은_이다(String symbol) {
+    @Then("기축통화 잠금 잔고는 {bigdecimal}이다")
+    public void 기축통화_잠금_잔고는_이다(BigDecimal amount) {
         apiClient
                 .getLastResponse()
                 .expectBody()
-                .jsonPath("$.data.baseCurrencySymbol")
-                .isEqualTo(symbol);
+                .jsonPath("$.data.baseCurrencyLocked")
+                .value(
+                        value ->
+                                assertThat(new BigDecimal(value.toString()).compareTo(amount))
+                                        .isZero());
     }
 
-    @Then("기축통화 잔고는 {long}이다")
-    public void 기축통화_잔고는_이다(Long balance) {
+    @Then("코인 잔고 개수는 {int}개이다")
+    public void 코인_잔고_개수는_개이다(int count) {
         apiClient
                 .getLastResponse()
                 .expectBody()
-                .jsonPath("$.data.baseCurrencyBalance")
-                .isEqualTo(balance);
-    }
-
-    @Then("보유 코인 개수는 {int}개이다")
-    public void 보유_코인_개수는_개이다(int count) {
-        apiClient
-                .getLastResponse()
-                .expectBody()
-                .jsonPath("$.data.holdings.length()")
+                .jsonPath("$.data.balances.length()")
                 .isEqualTo(count);
     }
 
-    @Then("첫 번째 코인 심볼은 {string}이다")
-    public void 첫_번째_코인_심볼은_이다(String symbol) {
+    @Then("첫 번째 코인의 coinId는 {long}이다")
+    public void 첫_번째_코인의_coinId는_이다(Long coinId) {
         apiClient
                 .getLastResponse()
                 .expectBody()
-                .jsonPath("$.data.holdings[0].coinSymbol")
-                .isEqualTo(symbol);
+                .jsonPath("$.data.balances[0].coinId")
+                .isEqualTo(coinId);
     }
 
-    @Then("첫 번째 코인 현재가는 {long}이다")
-    public void 첫_번째_코인_현재가는_이다(Long price) {
+    @Then("첫 번째 코인의 사용 가능 잔고는 {bigdecimal}이다")
+    public void 첫_번째_코인의_사용_가능_잔고는_이다(BigDecimal amount) {
         apiClient
                 .getLastResponse()
                 .expectBody()
-                .jsonPath("$.data.holdings[0].currentPrice")
-                .isEqualTo(price);
+                .jsonPath("$.data.balances[0].available")
+                .value(
+                        value ->
+                                assertThat(new BigDecimal(value.toString()).compareTo(amount))
+                                        .isZero());
+    }
+
+    @Then("첫 번째 코인의 잠금 잔고는 {bigdecimal}이다")
+    public void 첫_번째_코인의_잠금_잔고는_이다(BigDecimal amount) {
+        apiClient
+                .getLastResponse()
+                .expectBody()
+                .jsonPath("$.data.balances[0].locked")
+                .value(
+                        value ->
+                                assertThat(new BigDecimal(value.toString()).compareTo(amount))
+                                        .isZero());
     }
 
     private void insertUsers() {
@@ -120,11 +115,7 @@ public class MyHoldingsStepDefinition {
                         + "(2, '트레이더2', true)");
     }
 
-    private void insertExchangeAndCoins() {
-        jdbcTemplate.execute(
-                "INSERT IGNORE INTO exchange_market (exchange_id, name, market_type,"
-                        + " base_currency_coin_id, fee_rate) VALUES (1, 'Upbit', 'DOMESTIC', 1,"
-                        + " 0.000500), (2, 'Bithumb', 'DOMESTIC', 1, 0.000500)");
+    private void insertCoins() {
         jdbcTemplate.execute(
                 "INSERT IGNORE INTO coin (coin_id, symbol, name) VALUES "
                         + "(1, 'KRW', '원화'), "
@@ -132,16 +123,22 @@ public class MyHoldingsStepDefinition {
                         + "(3, 'ETH', '이더리움')");
     }
 
-    private void insertExchangeCoinMappings() {
+    private void insertExchanges() {
+        jdbcTemplate.execute(
+                "INSERT IGNORE INTO exchange_market (exchange_id, name, market_type,"
+                        + " base_currency_coin_id, fee_rate) VALUES (1, 'Upbit', 'DOMESTIC', 1,"
+                        + " 0.000500), (2, 'Bithumb', 'DOMESTIC', 1, 0.000500)");
+    }
+
+    private void insertExchangeCoins() {
         jdbcTemplate.execute(
                 "INSERT IGNORE INTO exchange_coin (exchange_coin_id, exchange_id, coin_id) VALUES "
                         + "(10, 1, 2), "
                         + "(11, 1, 3)");
     }
 
-    private void insertInvestmentRound() {
+    private void insertInvestmentRounds() {
         LocalDateTime now = LocalDateTime.now();
-        jdbcTemplate.execute("DELETE FROM investment_round WHERE round_id IN (1, 2)");
         jdbcTemplate.update(
                 "INSERT INTO investment_round (round_id, user_id, round_number, initial_seed,"
                         + " emergency_funding_limit, emergency_charge_count, status, started_at,"
@@ -172,13 +169,12 @@ public class MyHoldingsStepDefinition {
 
     private void insertWallets() {
         LocalDateTime now = LocalDateTime.now();
-        jdbcTemplate.execute("DELETE FROM wallet WHERE wallet_id IN (1, 2)");
         jdbcTemplate.update(
                 "INSERT INTO wallet (wallet_id, round_id, exchange_id, seed_amount, created_at)"
                         + " VALUES (?, ?, ?, ?, ?)",
                 1L,
                 1L,
-                EXCHANGE_ID,
+                EXCHANGE_UPBIT_ID,
                 new BigDecimal("10000000.00000000"),
                 now);
         jdbcTemplate.update(
@@ -186,20 +182,35 @@ public class MyHoldingsStepDefinition {
                         + " VALUES (?, ?, ?, ?, ?)",
                 2L,
                 1L,
-                2L,
+                EXCHANGE_BITHUMB_ID,
                 new BigDecimal("5000000.00000000"),
                 now);
     }
 
     private void insertWalletBalances() {
-        jdbcTemplate.execute("DELETE FROM wallet_balance WHERE wallet_id IN (1, 2)");
+        // 지갑 1: KRW(기축통화) + BTC + ETH
         jdbcTemplate.update(
                 "INSERT INTO wallet_balance (wallet_id, coin_id, available, locked) VALUES (?, ?,"
                         + " ?, ?)",
                 1L,
                 KRW_COIN_ID,
                 new BigDecimal("2450000.00000000"),
+                new BigDecimal("150000.00000000"));
+        jdbcTemplate.update(
+                "INSERT INTO wallet_balance (wallet_id, coin_id, available, locked) VALUES (?, ?,"
+                        + " ?, ?)",
+                1L,
+                BTC_COIN_ID,
+                new BigDecimal("0.05234100"),
+                new BigDecimal("0.00100000"));
+        jdbcTemplate.update(
+                "INSERT INTO wallet_balance (wallet_id, coin_id, available, locked) VALUES (?, ?,"
+                        + " ?, ?)",
+                1L,
+                ETH_COIN_ID,
+                new BigDecimal("1.24500000"),
                 BigDecimal.ZERO);
+        // 지갑 2: KRW(기축통화)만 존재
         jdbcTemplate.update(
                 "INSERT INTO wallet_balance (wallet_id, coin_id, available, locked) VALUES (?, ?,"
                         + " ?, ?)",
@@ -207,17 +218,5 @@ public class MyHoldingsStepDefinition {
                 KRW_COIN_ID,
                 new BigDecimal("5000000.00000000"),
                 BigDecimal.ZERO);
-    }
-
-    private void setUpHoldings() {
-        holdingAdapter.setHolding(
-                1L, BTC_COIN_ID, new BigDecimal("132500000"), new BigDecimal("0.052341"), 0);
-        holdingAdapter.setHolding(
-                1L, ETH_COIN_ID, new BigDecimal("5120000"), new BigDecimal("1.245"), 0);
-    }
-
-    private void setUpLivePrices() {
-        livePriceAdapter.setPrice(BTC_EXCHANGE_COIN_ID, new BigDecimal("143250000"));
-        livePriceAdapter.setPrice(ETH_EXCHANGE_COIN_ID, new BigDecimal("4821000"));
     }
 }
